+++ services/learn-card-network/vc-api-service/CHANGELOG.md
@@ -1,18 +1,68 @@
 # @learncard/vc-api-service
 
+## 1.0.4
+
+### Patch Changes
+
+- Updated dependencies [[`75d1816d5bbbd17772ef4c1c6b4932deacb9ccb5`](https://github.com/learningeconomy/LearnCard/commit/75d1816d5bbbd17772ef4c1c6b4932deacb9ccb5), [`6315fa3346cf75df2c6cbabe78962a9faa408781`](https://github.com/learningeconomy/LearnCard/commit/6315fa3346cf75df2c6cbabe78962a9faa408781), [`80d2ebf54bb5a643808f8f7d908cf68758903dce`](https://github.com/learningeconomy/LearnCard/commit/80d2ebf54bb5a643808f8f7d908cf68758903dce), [`75d1816d5bbbd17772ef4c1c6b4932deacb9ccb5`](https://github.com/learningeconomy/LearnCard/commit/75d1816d5bbbd17772ef4c1c6b4932deacb9ccb5)]:
+    - @learncard/expiration-plugin@1.2.34
+    - @learncard/types@5.19.0
+    - @learncard/vc-plugin@1.6.0
+    - @learncard/core@9.4.34
+    - @learncard/didkey-plugin@1.1.34
+    - @learncard/didkit-plugin@1.9.14
+    - @learncard/encryption-plugin@1.1.34
+    - @learncard/learn-card-plugin@1.2.34
+    - @learncard/vc-templates-plugin@1.1.34
+    - @learncard/crypto-plugin@1.1.34
+    - @learncard/dynamic-loader-plugin@1.1.34
+
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`55284a9b1613d46491a03d2d527b192eb2db64c9`](https://github.com/learningeconomy/LearnCard/commit/55284a9b1613d46491a03d2d527b192eb2db64c9), [`55284a9b1613d46491a03d2d527b192eb2db64c9`](https://github.com/learningeconomy/LearnCard/commit/55284a9b1613d46491a03d2d527b192eb2db64c9)]:
+    - @learncard/didkit-plugin@1.9.13
+    - @learncard/core@9.4.33
+    - @learncard/crypto-plugin@1.1.33
+    - @learncard/didkey-plugin@1.1.33
+    - @learncard/dynamic-loader-plugin@1.1.33
+    - @learncard/encryption-plugin@1.1.33
+    - @learncard/expiration-plugin@1.2.33
+    - @learncard/learn-card-plugin@1.2.33
+    - @learncard/vc-plugin@1.5.13
+    - @learncard/vc-templates-plugin@1.1.33
+
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`127eec62c5756b6f737028fa2e6a5e2277dc647a`](https://github.com/learningeconomy/LearnCard/commit/127eec62c5756b6f737028fa2e6a5e2277dc647a), [`cb7aef7dfc028824e5093230c7e46d0376481e36`](https://github.com/learningeconomy/LearnCard/commit/cb7aef7dfc028824e5093230c7e46d0376481e36)]:
+    - @learncard/types@5.18.3
+    - @learncard/didkit-plugin@1.9.12
+    - @learncard/core@9.4.32
+    - @learncard/didkey-plugin@1.1.32
+    - @learncard/encryption-plugin@1.1.32
+    - @learncard/learn-card-plugin@1.2.32
+    - @learncard/vc-plugin@1.5.12
+    - @learncard/vc-templates-plugin@1.1.32
+    - @learncard/crypto-plugin@1.1.32
+    - @learncard/dynamic-loader-plugin@1.1.32
+    - @learncard/expiration-plugin@1.2.32
+
 ## 1.0.1
 
 ### Patch Changes
 
--   Updated dependencies [[`69102b041690d8d161379ce9343c50a3dec0c50e`](https://github.com/learningeconomy/LearnCard/commit/69102b041690d8d161379ce9343c50a3dec0c50e)]:
-    -   @learncard/types@5.18.2
-    -   @learncard/core@9.4.31
-    -   @learncard/didkey-plugin@1.1.31
-    -   @learncard/didkit-plugin@1.9.11
-    -   @learncard/encryption-plugin@1.1.31
-    -   @learncard/learn-card-plugin@1.2.31
-    -   @learncard/vc-plugin@1.5.11
-    -   @learncard/vc-templates-plugin@1.1.31
-    -   @learncard/crypto-plugin@1.1.31
-    -   @learncard/dynamic-loader-plugin@1.1.31
-    -   @learncard/expiration-plugin@1.2.31
+- Updated dependencies [[`69102b041690d8d161379ce9343c50a3dec0c50e`](https://github.com/learningeconomy/LearnCard/commit/69102b041690d8d161379ce9343c50a3dec0c50e)]:
+    - @learncard/types@5.18.2
+    - @learncard/core@9.4.31
+    - @learncard/didkey-plugin@1.1.31
+    - @learncard/didkit-plugin@1.9.11
+    - @learncard/encryption-plugin@1.1.31
+    - @learncard/learn-card-plugin@1.2.31
+    - @learncard/vc-plugin@1.5.11
+    - @learncard/vc-templates-plugin@1.1.31
+    - @learncard/crypto-plugin@1.1.31
+    - @learncard/dynamic-loader-plugin@1.1.31
+    - @learncard/expiration-plugin@1.2.31